go mod tidy. packages moved under pkg/
