+++ .gitignore
@@ -22,3 +22,4 @@ go.work
 .idea
 dist
 .venv/
+.repomix/
