+++ go.mod
@@ -1,13 +1,15 @@
 module github.com/yusufcanb/tlm
 
-go 1.21
+go 1.22
+
+toolchain go1.22.0
 
 require (
 	github.com/Masterminds/semver v1.5.0
 	github.com/charmbracelet/huh v0.3.0
 	github.com/charmbracelet/huh/spinner v0.0.0-20240209193029-45947515c4cf
 	github.com/charmbracelet/lipgloss v0.9.1
-	github.com/jmorganca/ollama v0.1.25
+	github.com/jmorganca/ollama v0.1.29
 	github.com/spf13/viper v1.18.2
 	github.com/urfave/cli/v2 v2.27.1
 )
@@ -46,7 +48,6 @@ require (
 	github.com/xrash/smetrics v0.0.0-20231213231151-1d8dd44e695e // indirect
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.9.0 // indirect
-	golang.org/x/crypto v0.16.0 // indirect
 	golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
 	golang.org/x/sync v0.6.0 // indirect
 	golang.org/x/sys v0.17.0 // indirect
