+++ pkg/app/app.go
@@ -3,13 +3,14 @@ package app
 import (
 	_ "embed"
 	"fmt"
-	ollama "github.com/jmorganca/ollama/api"
-	"github.com/yusufcanb/tlm/config"
-	"github.com/yusufcanb/tlm/explain"
-	"github.com/yusufcanb/tlm/install"
-	"github.com/yusufcanb/tlm/suggest"
 	"io/fs"
 	"runtime"
+
+	ollama "github.com/jmorganca/ollama/api"
+	"github.com/yusufcanb/tlm/pkg/config"
+	"github.com/yusufcanb/tlm/pkg/explain"
+	"github.com/yusufcanb/tlm/pkg/install"
+	"github.com/yusufcanb/tlm/pkg/suggest"
 
 	"github.com/urfave/cli/v2"
 )
