+++ pkg/config/cli.go
@@ -3,15 +3,17 @@ package config
 import (
 	"errors"
 	"fmt"
+	"net/url"
+
 	"github.com/spf13/viper"
 	"github.com/urfave/cli/v2"
-	"github.com/yusufcanb/tlm/shell"
-	"net/url"
+	"github.com/yusufcanb/tlm/pkg/shell"
 )
 
 var (
 	defaultLLMHost = "http://localhost:11434"
 	shellKey       = "shell"
+	llmModelKey    = "llm.model"
 	llmHostKey     = "llm.host"
 	llmExplainKey  = "llm.explain"
 	llmSuggestKey  = "llm.suggest"
@@ -22,6 +24,7 @@ func (c *Config) Action(_ *cli.Context) error {
 
 	form := ConfigForm{
 		host:    viper.GetString(llmHostKey),
+		model:   viper.GetString(llmModelKey),
 		shell:   viper.GetString(shellKey),
 		explain: viper.GetString(llmExplainKey),
 		suggest: viper.GetString(llmSuggestKey),
@@ -34,6 +37,7 @@ func (c *Config) Action(_ *cli.Context) error {
 
 	viper.Set(shellKey, form.shell)
 	viper.Set(llmHostKey, form.host)
+	viper.Set(llmModelKey, form.model)
 	viper.Set(llmExplainKey, form.explain)
 	viper.Set(llmSuggestKey, form.suggest)
