+++ pkg/config/form.go
@@ -6,6 +6,7 @@ type ConfigForm struct {
 	form *huh.Form
 
 	host    string
+	model   string
 	shell   string
 	explain string
 	suggest string
@@ -14,10 +15,20 @@ type ConfigForm struct {
 func (c *ConfigForm) Run() error {
 	c.form = huh.NewForm(
 		huh.NewGroup(
-
 			huh.NewInput().
 				Title("Ollama").
 				Value(&c.host),
+
+			huh.NewSelect[string]().
+				Title("Model").
+				Description("Overrides platform's shell for suggestions").
+				Options(
+					huh.NewOption("Automatic", "auto"),
+					huh.NewOption("Powershell (Windows)", "powershell"),
+					huh.NewOption("Bash (Linux)", "bash"),
+					huh.NewOption("Zsh (macOS)", "zsh"),
+				).
+				Value(&c.shell),
 
 			huh.NewSelect[string]().
 				Title("Shell").
