+++ pkg/explain/cli.go
@@ -3,9 +3,10 @@ package explain
 import (
 	"context"
 	"fmt"
-	"github.com/urfave/cli/v2"
-	"github.com/yusufcanb/tlm/shell"
 	"os"
+
+	"github.com/urfave/cli/v2"
+	"github.com/yusufcanb/tlm/pkg/shell"
 )
 
 func (e *Explain) before(_ *cli.Context) error {
@@ -16,25 +17,25 @@ func (e *Explain) before(_ *cli.Context) error {
 		os.Exit(-1)
 	}
 
-	list, err := e.api.List(context.Background())
+	// list, err := e.api.List(context.Background())
 	if err != nil {
 		fmt.Println(shell.Err() + " " + err.Error())
 		fmt.Println(shell.Err() + " Ollama connection failed. Please check your Ollama if it's running or configured correctly.")
 		os.Exit(-1)
 	}
 
-	found := false
-	for _, model := range list.Models {
-		if model.Name == e.tag {
-			found = true
-			break
-		}
-	}
+	// found := false
+	// for _, model := range list.Models {
+	// 	if model.Name == e.tag {
+	// 		found = true
+	// 		break
+	// 	}
+	// }
 
-	if !found {
-		fmt.Println(shell.Err() + " " + "tlm's explain model not found.\n\nPlease run `tlm deploy` to deploy tlm models first.")
-		os.Exit(-1)
-	}
+	// if !found {
+	// 	fmt.Println(shell.Err() + " " + "tlm's explain model not found.\n\nPlease run `tlm deploy` to deploy tlm models first.")
+	// 	os.Exit(-1)
+	// }
 
 	return nil
 }
