+++ pkg/shell/checkpoint.go
