+++ pkg/shell/errors.go
