+++ pkg/shell/shell_darwin.go
