+++ pkg/shell/shell_linux.go
