+++ pkg/shell/shell_windows.go
