+++ pkg/suggest/suggest_test.go
@@ -1,10 +1,11 @@
 package suggest_test
 
 import (
-	ollama "github.com/jmorganca/ollama/api"
-	"github.com/yusufcanb/tlm/config"
-	"github.com/yusufcanb/tlm/suggest"
 	"testing"
+
+	ollama "github.com/jmorganca/ollama/api"
+	"github.com/yusufcanb/tlm/pkg/config"
+	"github.com/yusufcanb/tlm/pkg/suggest"
 )
 
 func TestSuggest(t *testing.T) {
